Move protocol for recipe creator view model to separate file
